requirements.txt
aiogram==3.0.0
environs==9.5.0
fastapi==0.103.1
pip-chill==1.0.3
psycopg2-binary==2.9.7
pyjwt==2.8.0
python-jose==3.3.0
requests==2.31.0
sqlalchemy==2.0.20
uvicorn==0.23.2
python-decouple==3.8
